# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Authentication and Authorization API",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).For this sample, you can use the api key `special-key` to test the authorization     filters.",
        "termsOfService": "http://Lanthanion.com",
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/accesslogs/getallaccesslogs": {
            "post": {
                "tags": [
                    "access-logs-controller"
                ],
                "operationId": "getAllAccessLogs",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GeneralPagedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralPagedAccessLogsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/accesslogs/deleteaccesslog": {
            "delete": {
                "tags": [
                    "access-logs-controller"
                ],
                "operationId": "deleteAccessLog",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteAccessLogRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/accesslogs/getparticularaccesslog": {
            "post": {
                "tags": [
                    "access-logs-controller"
                ],
                "operationId": "getParticularAccessLog",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GetParticularAccessLogRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralAccessLogsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/address/getalladdresses": {
            "post": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "getAllAddresses",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GeneralPagedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralPagedAddressesResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/address/updateuseraddress": {
            "put": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "updateUserAddress",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserAddressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralAddressResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/address/deleteuseraddress": {
            "delete": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "deleteUserAddress",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteUserAddressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/address/getparticularuseraddress": {
            "post": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "getParticularUserAddress",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GetParticularUserAddressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralAddressResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GeneralPagedRequest": {
                "required": [
                    "isDeleted",
                    "items",
                    "page"
                ],
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                }
            },
            "AccessLogs": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "deviceId": {
                        "type": "string"
                    },
                    "deviceIp": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "browserAgent": {
                        "type": "string"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Clients"
                    },
                    "accessToken": {
                        "$ref": "#/components/schemas/Token"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Clients": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "clientId": {
                        "type": "string"
                    },
                    "clientSecret": {
                        "type": "string"
                    },
                    "nickName": {
                        "type": "string"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "GeneralPagedAccessLogsResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccessLogs"
                        }
                    }
                }
            },
            "Token": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "accesstoken": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "ttl": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Users": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "isLoggedIn": {
                        "type": "boolean"
                    },
                    "userType": {
                        "type": "string",
                        "enum": [
                            "ADVERTISER",
                            "AD_OWNER",
                            "AFFILIATE",
                            "ADMIN"
                        ]
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "DeleteAccessLogRequest": {
                "required": [
                    "accessLogId"
                ],
                "type": "object",
                "properties": {
                    "accessLogId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GeneralResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "GetParticularAccessLogRequest": {
                "required": [
                    "accessLogId"
                ],
                "type": "object",
                "properties": {
                    "accessLogId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GeneralAccessLogsResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AccessLogs"
                    }
                }
            },
            "Address": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "street": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "GeneralPagedAddressesResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Address"
                        }
                    }
                }
            },
            "UpdateUserAddressRequest": {
                "required": [
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "city": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "street": {
                        "type": "string"
                    }
                }
            },
            "GeneralAddressResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Address"
                    }
                }
            },
            "DeleteUserAddressRequest": {
                "required": [
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GetParticularUserAddressRequest": {
                "required": [
                    "isDeleted",
                    "user"
                ],
                "type": "object",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "API KEY / API SECRET": {
                "type": "apiKey",
                "scheme": "apiKey"
            }
        }
    }
}